Persist Theme & View
Currently there's a flaw with the implementation, and that's the fact
that the UI will be rebuilt upon fetching from SharedPreferences.

This means that users will see a flash of default theme (darkTheme)
before seeing what their saved theme is.

The reason for this is because I am setting the default theme via
`setTheme`, but the default theme is, as shown here:

`ThemeManager() : super(darkTheme);`

The roadblock here now is that I am unsure how to pass in default theme
to StateNotifier, via Riverpod's StateNotifierProvider, so I had to hardcode the default to `darkTheme`.

If anyone has got any idea how to fix this, PRs are welcomed.