Improve CommentTile indent for better visibility
This commit changes the comment from a single indent color to multiple
(up to their current depth).

This makes it much easier to trace the comments when it's a long-running
thread.

This internally uses a hackish way to achieve this effect, using Stack &
Positioned. The reason is because we need the Container of width 2 to
expand its height fully, but it's not possible in a Row widget.

The only other option I found was IntrinsicHeight. While that solution
looks much cleaner, there are a few drawbacks:
- IntrinsicHeight is warned to be expensive
- For some reason, using IntrinsicHeight messes with the constraint for
Html, so the HTML gets cut off.